generating password-validator.test
