+++ Sprint-3/revise/implement/password-validator.test.js
@@ -1,45 +1,44 @@
+const isValidPassword = require('./password-validator'); // Adjust the path as needed
+
+describe('isValidPassword', () => {
+  // Sample list of previous passwords
+  const previousPasswords = ["password123", "abc123", "welcome1"];
+
+  test('returns false if the password is less than 5 characters long', () => {
+    expect(isValidPassword('1234', previousPasswords)).toBe(false);
+  });
+
+  test('returns false if the password does not contain an uppercase letter', () => {
+    expect(isValidPassword('password1!', previousPasswords)).toBe(false);
+  });
+
+  test('returns false if the password does not contain a lowercase letter', () => {
+    expect(isValidPassword('PASSWORD1!', previousPasswords)).toBe(false);
+  });
+
+  test('returns false if the password does not contain a number', () => {
+    expect(isValidPassword('Password!', previousPasswords)).toBe(false);
+  });
+
+  test('returns false if the password does not contain a special character', () => {
+    expect(isValidPassword('Password1', previousPasswords)).toBe(false);
+  });
+
+  test('returns false if the password is in the list of previous passwords', () => {
+    expect(isValidPassword('password123', previousPasswords)).toBe(false);
+    expect(isValidPassword('abc123', previousPasswords)).toBe(false);
+  });
+
+  test('returns true for a valid password', () => {
+    expect(isValidPassword('Password1!', previousPasswords)).toBe(true);
+  });
+
+  test('returns false for a valid password format but already used', () => {
+    expect(isValidPassword('welcome1', previousPasswords)).toBe(false);
+  });
+});
 
 
-function isValidPassword(password, previousPasswords) {
-    // 1. Check length
-    if (password.length < 5) {
-        return false;
-    }
-    
-    // 2. Check for at least one uppercase letter
-    if (!/[A-Z]/.test(password)) {
-        return false;
-    }
 
-    // 3. Check for at least one lowercase letter
-    if (!/[a-z]/.test(password)) {
-        return false;
-    }
 
-    // 4. Check for at least one number
-    if (!/[0-9]/.test(password)) {
-        return false;
-    }
-
-    // 5. Check for at least one special character
-    if (!/[!#$%.*&]/.test(password)) {
-        return false;
-    }
-
-    // 6. Check if the password is not in the list of previous passwords
-    if (previousPasswords.includes(password)) {
-        return false;
-    }
-
-    return true;
-}
-
-// Sample list of previous passwords
-const previousPasswords = ["password123", "abc123", "welcome1"];
-
-// Test cases
-console.log(isValidPassword("abc123!", previousPasswords));  // false
-console.log(isValidPassword("abc123", previousPasswords));   // false 
-console.log(isValidPassword("ABC123!", previousPasswords));  // false 
-console.log(isValidPassword("password123", previousPasswords)); // false
